+++ CHANGELOG.md
@@ -4,5 +4,5 @@
 
 ### New features
 
-- New parsers `find_byte` and `find_byte_keep` used for searching for
-  bytes in arch native chunk sizes (e.g. 8 bytes on 64-bit).
+- New parser `find_byte` used for searching for bytes in arch native
+  chunk sizes (e.g. 8 bytes on 64-bit).